+++ Runtime/Scripts/Coroutines/FrameDelayedAction.cs
@@ -3,8 +3,13 @@ using UnityEngine;
 
 namespace BeltainsTools.Coroutines
 {
-    public class FrameDelayedAction //Should get this to work without a specified executor (On a beltainsTools monobehaviour maybe) to avoid "can't run coroutine because object is inactive" bullshit
+    public class FrameDelayedAction
     {
+        public static void Execute(System.Action action)
+        {
+            CoroutineRunner.Run(ExecuteActionAfterFrame(action));
+        }
+
         public static void Execute(System.Action action, MonoBehaviour executor)
         {
             executor.StartCoroutine(ExecuteActionAfterFrame(action));